+++ src/lib.rs
@@ -12,6 +12,7 @@ pub mod rfc1870;
 pub mod rfc3207;
 pub mod rfc3461;
 pub mod rfc3463;
+pub mod rfc4505;
 pub mod rfc4616;
 pub mod rfc4954;
 pub mod rfc5321;